Added Lists

+++ app/src/main/java/com/emerycprimeau/gameq/models/CurrentUser.java
@@ -1,6 +1,6 @@
 package com.emerycprimeau.gameq.models;
 
 public class CurrentUser {
-    public static long currentId;
+    public static int currentId;
     public static String email;
 }
