+++ app/src/main/java/com/emerycprimeau/gameq/models/transfer/GameRequest.java
@@ -1,5 +1,5 @@
 package com.emerycprimeau.gameq.models.transfer;
 
 public class GameRequest {
-    public Long userId;
+    public int userId;
 }
